// src/lib/api/types/updateAccountPasswordRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Adisu Backend API
 * Backend API for the Adisu platform
 * OpenAPI spec version: 1.0.0
 */

export interface UpdateAccountPasswordRequest {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  /**
   * Password (min 8 chars, 1 uppercase, 1 lowercase, 1 digit)
   * @minLength 8
   * @maxLength 128
   */
  confirmPassword: string;
  /**
   * Password
   * @minLength 1
   * @maxLength 128
   */
  existingPassword: string;
  /**
   * Password (min 8 chars, 1 uppercase, 1 lowercase, 1 digit)
   * @minLength 8
   * @maxLength 128
   */
  newPassword: string;
}